+++ lib/src/tokens/geometry/form_factor.dart
@@ -1,8 +1,0 @@
-/// Identifiers for device form factors.
-library;
-
-enum MaterialFormFactor {
-  small,
-  medium,
-  large,
-}